backend/requirements.txt
# Core Django Framework
Django==5.2.2

# Django REST Framework for API development
djangorestframework==3.15.2

# Django Filter for API filtering
django-filter==23.5

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# PostgreSQL Database Adapter
psycopg2-binary==2.9.9

# CORS Headers for frontend integration
django-cors-headers==4.3.1

# Environment variable management
python-decouple==3.8

# API Documentation
drf-yasg==1.21.7

# Password validation and security
django-extensions==3.2.3

# Security enhancements
django-ratelimit==4.1.0
django-security==0.12.0
django-csp==3.7

# HTML sanitization for XSS prevention
bleach==6.1.0

# Testing and development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
mixer==7.2.2

# Code quality
flake8==6.1.0
black==23.11.0

# Production server
gunicorn==21.2.0

# Static file handling
whitenoise==6.6.0

# AWS SDK for S3 and other AWS services
boto3==1.34.0
botocore==1.34.0

# Redis for caching and sessions
redis==5.0.1
django-redis==5.4.0

# Health checks
django-health-check==3.17.0

# Logging and monitoring
sentry-sdk==1.38.0

# Database connection pooling handled by Django's CONN_MAX_AGE setting
# See settings_production.py for configuration